reinit project

+++ recruitr/db/migrate/20191210205601_create_available_positions.rb
@@ -2,7 +2,7 @@ class CreateAvailablePositions < ActiveRecord::Migration[6.0]
   def change
     create_table :available_positions do |t|
       t.string :title
-      t.numeric :number_of_applicants
+      t.integer :number_of_applicants
       t.timestamps
     end
   end
